requirements.txt
Flask==3.0.3
boto3==1.34.131
psycopg[binary]
google-generativeai==0.8.5
gunicorn==22.0.0